:sparkles: 가게 상세 정보 > 탭바의 고정 시점을 감지하여 앱바 변경하는 기능 구현
